requirements.txt
streamlit
torch
torchvision
transformers
accelerate
pillow
numpy
sentencepiece
timm
safetensors
